Fix inconsistent count results after cardinality annotation (#57)
Annotating the queryset by the mapbox_context cardinality left us
with a row for each product/location combination.

(e.g.: if a product was linked to two locations with different
context lengths, that would result in a duplicated row).

What we really want to do is annotating the queryset with the length
of the longest mapbox context (i.e. most specific location) for ordering
purposes. This results in no duplication and consistent API results.