[PKB-81] Admin panel location add view (#121)
* Specify the place type for a location in admin panel's representation

There are several examples of a place which is also a region
(like London is both the city and the district). This
helps out disambiguate when adding new locations to a product.

* Override new location form in admin panel

To allow admin panel users to input a new location only by selecting
a matching location from mapbox.

* Mark locations added via admin panel as approved

* Style update

* Approve locations via adding

if they already exist but are not yet approved.